requirements.txt
Flask
flask_sqlalchemy
psycopg2-binary
python-dotenv
requests
pyjwt
livekit
gunicorn
